[16.0][IMP] budget: remove unused execution status report components
- Remove budget_execution_status_report and budget_execution_status_line models
- Remove budget_appropriation_report model and integrate hierarchy logic into overview report
- Remove unused budget_availability_widget component
- Remove interactive execution report components and views
- Clean up manifest, security rules, and template references
- Consolidate hierarchy building methods in budget_appropriation_overview_report
