+++ budget/models/__init__.py
@@ -2,16 +2,12 @@ from . import budget_account
 from . import budget_move
 from . import budget_move_line
 from . import budget_journal
-from . import budget_appropriation_report
 from . import budget_appropriation_overview_report
 
 # Modern budget models (renamed from _new to normal)
 from . import budget_commitment_mixin
 from . import budget_commitment
 
-# Budget execution status report
-from . import budget_execution_status_report
-from . import budget_execution_status_line
 
 # Budget integration components
 from . import budget_mixin
