+++ budget/security/ir.model.access.csv
@@ -5,10 +5,7 @@ access_budget_move_line_manager,budget_move_line_manager,model_budget_move_line,
 access_budget_account_user,budget_account_user,model_budget_account,,1,0,0,0
 access_budget_account_manager,budget_account_manager,model_budget_account,,1,1,1,1
 access_budget_journal_manager,budget_journal_manager,model_budget_journal,,1,1,1,1
-access_budget_appropriation_report_manager,budget_appropriation_report_manager,model_budget_appropriation_report,,1,1,1,1
 access_budget_appropriation_overview_report_manager,budget_appropriation_overview_report_manager,model_budget_appropriation_overview_report,,1,1,1,1
-access_budget_execution_status_report_manager,budget_execution_status_report_manager,model_budget_execution_status_report,,1,1,1,1
-access_budget_execution_status_line_manager,budget_execution_status_line_manager,model_budget_execution_status_line,,1,1,1,1
 access_budget_transfer_user,budget_transfer_user,model_budget_transfer,group_budget_transfer_user,1,1,1,0
 access_budget_transfer_approver,budget_transfer_approver,model_budget_transfer,group_budget_transfer_approver,1,1,1,0
 access_budget_transfer_manager,budget_transfer_manager,model_budget_transfer,group_budget_transfer_manager,1,1,1,1
