--- budget/static/src/components/budget_availability_widget/budget_availability_widget.scss
@@ -1,70 +1,0 @@
-.o_budget_availability_widget {
-    .card {
-        border: 1px solid #dee2e6;
-        box-shadow: 0 1px 3px rgba(0,0,0,0.05);
-        
-        .card-header {
-            background-color: #f8f9fa;
-            border-bottom: 1px solid #dee2e6;
-            padding: 0.75rem 1rem;
-            
-            h6 {
-                color: #495057;
-                font-weight: 600;
-            }
-        }
-        
-        .card-body {
-            padding: 1rem;
-        }
-    }
-    
-    .table-borderless {
-        td {
-            padding: 0.25rem 0.5rem;
-            
-            &.text-muted {
-                font-size: 0.875rem;
-            }
-            
-            &.font-weight-bold {
-                font-size: 1rem;
-            }
-        }
-    }
-    
-    .progress {
-        background-color: #e9ecef;
-        
-        .progress-bar {
-            font-weight: 600;
-            font-size: 0.875rem;
-        }
-    }
-    
-    .badge {
-        font-size: 0.75rem;
-        padding: 0.25rem 0.5rem;
-        
-        &.badge-success {
-            background-color: #28a745;
-        }
-        
-        &.badge-warning {
-            background-color: #ffc107;
-            color: #212529;
-        }
-        
-        &.badge-danger {
-            background-color: #dc3545;
-        }
-    }
-    
-    .alert {
-        font-size: 0.875rem;
-        
-        i {
-            margin-right: 0.5rem;
-        }
-    }
-}
