--- budget/static/src/components/budget_availability_widget/budget_availability_widget.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates xml:space="preserve">
-    <t t-name="budget.BudgetAvailabilityWidget" owl="1">
-        <div class="o_budget_availability_widget">
-            <t t-if="state.loading">
-                <div class="text-center p-3">
-                    <i class="fa fa-spinner fa-spin"/> Loading budget data...
-                </div>
-            </t>
-            <t t-elif="state.error">
-                <div class="alert alert-danger m-2">
-                    <i class="fa fa-exclamation-triangle"/> <t t-esc="state.error"/>
-                </div>
-            </t>
-            <t t-elif="state.data">
-                <div class="card m-2">
-                    <div class="card-header">
-                        <h6 class="mb-0">
-                            <i class="fa fa-chart-pie"/> Budget Availability Analysis
-                        </h6>
-                    </div>
-                    <div class="card-body">
-                        <div class="row">
-                            <div class="col-md-6">
-                                <table class="table table-sm table-borderless">
-                                    <tr>
-                                        <td class="text-muted">Appropriated:</td>
-                                        <td class="text-end font-weight-bold">
-                                            <t t-esc="formatAmount(state.data.appropriated)"/>
-                                        </td>
-                                    </tr>
-                                    <tr>
-                                        <td class="text-muted">Reserved:</td>
-                                        <td class="text-end text-warning">
-                                            - <t t-esc="formatAmount(state.data.reserved)"/>
-                                        </td>
-                                    </tr>
-                                    <tr>
-                                        <td class="text-muted">Consumed:</td>
-                                        <td class="text-end text-danger">
-                                            - <t t-esc="formatAmount(state.data.consumed)"/>
-                                        </td>
-                                    </tr>
-                                    <tr class="border-top">
-                                        <td class="text-muted">Available:</td>
-                                        <td class="text-end font-weight-bold" t-att-class="statusClass">
-                                            <t t-esc="formatAmount(state.data.available)"/>
-                                        </td>
-                                    </tr>
-                                </table>
-                            </div>
-                            <div class="col-md-6 d-flex align-items-center justify-content-center">
-                                <div class="text-center">
-                                    <div class="progress" style="height: 30px; width: 200px;">
-                                        <div class="progress-bar" 
-                                             t-att-class="'bg-' + (availablePercentage >= 50 ? 'success' : availablePercentage >= 20 ? 'warning' : 'danger')"
-                                             role="progressbar" 
-                                             t-att-style="'width: ' + availablePercentage + '%'"
-                                             t-att-aria-valuenow="availablePercentage"
-                                             aria-valuemin="0" 
-                                             aria-valuemax="100">
-                                            <t t-esc="Math.round(availablePercentage)"/>% Available
-                                        </div>
-                                    </div>
-                                    <small class="text-muted mt-2 d-block">
-                                        Budget Utilization
-                                    </small>
-                                </div>
-                            </div>
-                        </div>
-                        
-                        <t t-if="state.data.details and state.data.details.length">
-                            <div class="mt-3">
-                                <h6 class="text-muted">Recent Transactions</h6>
-                                <table class="table table-sm">
-                                    <thead>
-                                        <tr>
-                                            <th>Date</th>
-                                            <th>Type</th>
-                                            <th>Reference</th>
-                                            <th class="text-end">Amount</th>
-                                        </tr>
-                                    </thead>
-                                    <tbody>
-                                        <t t-foreach="state.data.details.slice(0, 5)" t-as="detail" t-key="detail.id">
-                                            <tr>
-                                                <td><t t-esc="detail.date"/></td>
-                                                <td>
-                                                    <span t-att-class="'badge badge-' + (detail.type === 'appropriation' ? 'success' : detail.type === 'reserved' ? 'warning' : 'danger')">
-                                                        <t t-esc="detail.type"/>
-                                                    </span>
-                                                </td>
-                                                <td><t t-esc="detail.reference"/></td>
-                                                <td class="text-end">
-                                                    <t t-if="detail.type === 'appropriation'">+</t>
-                                                    <t t-else="">-</t>
-                                                    <t t-esc="formatAmount(Math.abs(detail.amount))"/>
-                                                </td>
-                                            </tr>
-                                        </t>
-                                    </tbody>
-                                </table>
-                            </div>
-                        </t>
-                    </div>
-                </div>
-            </t>
-            <t t-else="">
-                <div class="alert alert-info m-2">
-                    <i class="fa fa-info-circle"/> Select budget account and analytic dimensions to view availability
-                </div>
-            </t>
-        </div>
-    </t>
-</templates>
