--- budget/views/budget_execution_report_interactive_views.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <menuitem
-        id="menu_budget_execution_report_interactive"
-        name="Budget Execution Report (Interactive)"
-        parent="budget_reporting_menu"
-        action="action_budget_execution_report_interactive"
-        sequence="25"
-    />
-</odoo>
